+++ src/main/java/com/awbd/ecommerce/service/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package com.awbd.ecommerce.service;
 
 import com.awbd.ecommerce.dto.ProductDTO;
 import com.awbd.ecommerce.dto.ReviewDTO;
+import com.awbd.ecommerce.exception.ResourceNotFoundException;
 import com.awbd.ecommerce.helper.BeanHelper;
 import com.awbd.ecommerce.model.Product;
 import com.awbd.ecommerce.model.Review;
@@ -11,7 +12,6 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,23 +41,26 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductDTO findById(Long l) {
-        Optional<Product> productOptional = productRepository.findById(l);
-        if (productOptional.isEmpty()) {
-            throw new RuntimeException("Product not found!");
-        }
-        return modelMapper.map(productOptional.get(), ProductDTO.class);
+    public ProductDTO findById(Long id) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Product with id" + id + " not found!"));
+
+        return modelMapper.map(product, ProductDTO.class);
     }
 
     @Override
     public void deleteById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product with id " + id + " not found!");
+        }
+
         productRepository.deleteById(id);
     }
 
     @Override
     public ProductDTO update(Long id, ProductDTO productDTO) {
         Product product = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Order with id " + id + "doesn't exists"));
+                .orElseThrow(() -> new ResourceNotFoundException("Product with id " + id + " not found!"));
 
         BeanUtils.copyProperties(productDTO, product, BeanHelper.getNullPropertyNames(productDTO));
